+++ spec/app/db/schema.rb
@@ -11,12 +11,15 @@
 # It's strongly recommended that you check this file into your version control system.
 
 ActiveRecord::Schema[7.1].define(version: 2024_07_07_052907) do
+  # These are extensions that must be enabled in order to support this database
+  enable_extension "plpgsql"
+
   create_table "posts", force: :cascade do |t|
     t.string "title"
     t.integer "category"
     t.text "body"
     t.datetime "published_at"
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["user_id"], name: "index_posts_on_user_id"
@@ -26,7 +29,8 @@ ActiveRecord::Schema[7.1].define(version: 2024_07_07_052907) do
     t.string "name", null: false
     t.string "username", null: false
     t.boolean "active", default: false, null: false
-    t.integer "invitor_id"
+    t.bigint "invitor_id"
+    t.integer "role", default: 0, null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["invitor_id"], name: "index_users_on_invitor_id"