+++ lib/typelizer/serializer_plugins/ams.rb
@@ -3,19 +3,6 @@ require_relative "base"
 module Typelizer
   module SerializerPlugins
     class AMS < Base
-      def methods_to_typelize
-        [
-          :has_many, :has_one, :belongs_to,
-          :attribute, :attributes
-        ]
-      end
-
-      def typelize_method_transform(method:, name:, binding:, type:, attrs:)
-        return {binding.local_variable_get(:attr) => [type, attrs]} if method == :attribute
-
-        super
-      end
-
       def properties
         serializer._attributes_data.merge(serializer._reflections).flat_map do |key, association|
           type = association.options[:serializer] ? context.interface_for(association.options[:serializer]) : nil